requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
Flask-PyMongo==3.0.1
graphql-core==3.2.6
itsdangerous==2.2.0
Jinja2==3.1.6
lia-web==0.2.3
MarkupSafe==3.0.2
packaging==25.0
pymongo==4.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
strawberry-graphql==0.282.0
typing_extensions==4.15.0
Werkzeug==3.1.3
